backend_django/requirements.txt
# Django Core
Django==5.1.2
djangorestframework==3.15.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# Environment
django-environ==0.11.2

# Telegram
python-telegram-bot==21.6

# Async tasks (для будущего)
celery==5.3.6
redis==5.0.1

# WSGI Server
gunicorn==21.2.0

# Utilities
python-dateutil==2.8.2

